Update Department.cs
fixed

+++ src/UMS.Domain/Entities/Education/Department.cs
@@ -2,7 +2,7 @@
 {
     public class Department : BaseEntity
     {
-        public int FacultyID { get; set; }
+        public long FacultyID { get; set; }
         public string Name { get; set; }
     }
 }
